feat: add update feature for connected users which listing is theirs

+++ backend/controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require ('../models/Book');
+const fs = require('fs');
 
 // Create a new book in the database
 exports.createBook = (req, res, next) => {
@@ -33,6 +34,7 @@ exports.getAllBooks = (req, res, next) => {
   
   // Return the single book with the provided _id
 exports.getOneBook = (req, res, next) => {
+    console.log(req.params.id)
     Book.findOne({_id: req.params.id})
     .then((book) => {
         res.status(200).json(book);
@@ -43,27 +45,44 @@ exports.getOneBook = (req, res, next) => {
   };
 
   // Update the book with the provided  _id  with the data provided in the request body.
-exports.updateBook = (req, res, next) => {
-    const book = new Book({
-        _id: req.params.id,
-        ...req.body
-    });
-    Book.updateOne({_id: req.params.id}, book)
-    .then(() => {
-        res.status(201).json({message: 'Modified!'});
-    })
-    .catch((error) => {
-        res.status(400).json({ error });
-    });
-  };
+  exports.updateBook = (req, res, next) => {
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+  
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Not authorized'});
+            } else {
+                Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+                .then(() => res.status(200).json({message : 'Livre modifié !'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+ };
 
   // Delete the single book with the provided _id
   exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({_id: req.params.id})
-    .then(() => {
-        res.status(201).json({message: 'Deleted!'});
-    })
-    .catch((error) => {
-        res.status(404).json({ error });
-    });
-  };
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
